+++ frontend/src/App.jsx
@@ -2,12 +2,14 @@ import { useState } from 'react'
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
 import './App.css'
 
-import Navbar from './navbar/Navbar'
-import Dash from './views/Dash'
-import Instruments from './views/Instruments'
-import Packs from './views/Packs'
-import Procedures from './views/Procedures'
-
+import Navbar from './components/navbar/Navbar'
+import Dash from './components/views/Dash'
+import Instruments from './components/instruments/Instruments'
+import Instrument from './components/instruments/Instrument'
+import Packs from './components/packs/Packs'
+import Procedures from './components/procedures/Procedures'
+import Procedure from './components/procedures/Procedure'
+import ProcedureEdit from './components/procedures/ProcedureEdit'
 
 function App() {
   const [count, setCount] = useState(0)
@@ -21,6 +23,8 @@ function App() {
           <Route path="/instruments" element={<Instruments />} />
           <Route path="/packs" element={<Packs />} />
           <Route path="/procedures" element={<Procedures />} />
+          <Route path="/procedures/:id" element={<Procedure />} />
+          <Route path="/procedures/:id/edit" element={<ProcedureEdit />} />
         </Routes>
       </Router>
     </>