+++ frontend/src/components/instruments/Instrument.jsx
@@ -1,0 +1,10 @@
+import React from 'react'
+
+function Instrument() {
+    
+    return (
+        <div>Instrument</div>
+    )
+}
+
+export default Instrument