+++ frontend/src/components/packs/Pack.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function Pack() {
+    return (
+        <div>Pack</div>
+    )
+}
+
+export default Pack