+++ frontend/src/components/views/Dash.jsx
@@ -3,8 +3,7 @@ import React from 'react'
 function Dash() {
     return (
         <>
-        <h1>Dashboard.</h1>
-        <div>Dash</div>
+        <h1>Dashboard</h1>
         </>
     )
 }