+++ Changes.md
@@ -1,8 +1,11 @@
 #Frontend
     - [ ] Everything.
-    - [ ] import axios, react-router-dom
     - [ x ] Basic calls for PPI.
       - [ ] Change display to table for all three
+          - [ x ] Instruments
+              - [ ] Add search functionality
+              - [ ] Implement pagination
+              - [ ] Implement Filter & Actions 
         - [ ] Implement Search functionality.
 
 
@@ -17,3 +20,4 @@
 
     9/23:
     Added onHand to inventory, notes to Packs using Alembic.
+    - [ ] Consider backend pagination for instruments. Expecting ~1,039,000 characters, which may mean long loading times, but that is for the future.